requirements.txt
pillow
raylib
rich
st7789
